# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /encounters:
    get:
      operationId: listEncounters
      tags: [encounters]
      summary: List encounters in the active organization
      description: |
        Returns encounters scoped to the caller's active organization,
        ordered by createdAt desc. Optional patientId narrows to a
        single chart. Cap at 200 rows; revisit with cursor pagination if
        a busy clinic exceeds it.
      parameters:
        - name: patientId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: List of encounters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEncounters200"
        "401":
          description: Unauthenticated
        "409":
          description: Session has no active organization
    post:
      operationId: createEncounter
      tags: [encounters]
      summary: Start a new encounter
      description: |
        Patient must belong to the active organization. visitType='custom'
        requires customLabel; any other visitType forbids it. Defaults
        providerId to the current user when not supplied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEncounterRequest"
      responses:
        "201":
          description: Created encounter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Encounter" }
        "400":
          description: Invalid request or custom_label mismatch
        "404":
          description: Patient not found in the active organization
  /encounters/{id}:
    get:
      operationId: getEncounter
      tags: [encounters]
      summary: Read a single encounter
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Encounter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Encounter" }
        "404":
          description: Encounter not found in the active organization
    patch:
      operationId: updateEncounter
      tags: [encounters]
      summary: Update encounter fields and/or transition status
      description: |
        Status transitions are enforced server-side: scheduled →
        in_progress → completed are the legal forward steps; cancellation
        is allowed from any non-terminal state. Completed and cancelled
        are terminal. Auto-stamps startedAt / completedAt timestamps on
        the matching transition.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEncounterRequest"
      responses:
        "200":
          description: Updated encounter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Encounter" }
        "400":
          description: Invalid request
        "404":
          description: Encounter not found in the active organization
        "409":
          description: Illegal status transition
...
  /encounters/{id}/audit-timeline:
    get:
      operationId: listEncounterAuditTimeline
      tags: [encounters]
      summary: Encounter-scoped audit drilldown — every audit_log event tied to this encounter
      description: >-
        Returns audit_log rows for every resource attached to the
        encounter: the encounter itself, its note(s), coding sessions,
        billing suggestions, problem-list suggestions, and approved
        billing codes. Reverse-chronological, capped at the limit param.
        Non-admin — providers + billers see the same view (org-scoped).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      # `?limit=` is supported by the handler (default 200, capped at 500)
      # but kept out of the spec to dodge an Orval codegen name-clash on
      # mixed path + query params. Spec it formally when the codegen
      # config is loosened.
      responses:
        "200":
          description: Audit timeline
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/EncounterAuditEvent" }
        "404":
          description: Encounter not found in the active organization
...
components:
  schemas:
...
    VisitType:
      type: string
      enum:
        - new_patient
        - established_patient
        - follow_up
        - annual_physical
        - hospital_follow_up
        - procedure
        - telehealth
        - nursing_facility
        - custom
    EncounterStatus:
      type: string
      enum:
        - scheduled
        - in_progress
        - completed
        - cancelled
...
    Encounter:
      type: object
      required:
        - id
        - organizationId
        - patientId
        - visitType
        - status
        - isTelehealth
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        organizationId: { type: string }
        patientId: { type: string }
        providerId: { type: string, nullable: true }
        visitType: { $ref: "#/components/schemas/VisitType" }
        customLabel: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/EncounterStatus" }
        isTelehealth: { type: boolean }
        location: { type: string, nullable: true }
        scheduledAt: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        ehrEncounterRef:
          type: string
          nullable: true
          description: >-
            FHIR-style upstream encounter reference ("Encounter/<id>") set when
            this local encounter mirrors an Athena (or other EHR) chart entry.
            Required for real-mode chart-API writeback; null means the encounter
            isn't linked to the EHR yet and pushes will fail with a clear
            "not linked" message.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EncounterAuditEvent:
      type: object
      required: [id, at, action, resourceType]
      properties:
        id: { type: string }
        at:
          type: string
          format: date-time
        userId:
          type: ["string", "null"]
        userDisplayName:
          type: ["string", "null"]
        action: { type: string }
        resourceType: { type: string }
        resourceId:
          type: ["string", "null"]
        metadata:
          type: ["object", "null"]
          additionalProperties: true
    ListEncounters200:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Encounter" }
    CreateEncounterRequest:
      type: object
      required: [patientId, visitType]
      properties:
        patientId: { type: string, minLength: 1 }
        visitType: { $ref: "#/components/schemas/VisitType" }
        customLabel: { type: string, minLength: 1, maxLength: 120 }
        isTelehealth: { type: boolean }
        location: { type: string, maxLength: 120 }
        scheduledAt: { type: string, format: date-time }
        providerId: { type: string, minLength: 1 }
        ehrEncounterRef:
          type: string
          minLength: 1
          maxLength: 120
          description: >-
            Optional Athena (or other EHR) Encounter id at create time. Accepts
            "Encounter/<id>" or bare "<id>"; normalized server-side to
            "Encounter/<id>".
    UpdateEncounterRequest:
      type: object
      description: |
        All fields optional. Set a field to null to clear it where the
        column allows null (location, customLabel, scheduledAt,
        providerId). visitType / status / isTelehealth are non-nullable
        so null isn't accepted there.
      properties:
        visitType: { $ref: "#/components/schemas/VisitType" }
        customLabel: { type: string, minLength: 1, maxLength: 120, nullable: true }
        isTelehealth: { type: boolean }
        location: { type: string, maxLength: 120, nullable: true }
        status: { $ref: "#/components/schemas/EncounterStatus" }
        scheduledAt: { type: string, format: date-time, nullable: true }
        providerId: { type: string, minLength: 1, nullable: true }
        ehrEncounterRef:
          type: string
          minLength: 1
          maxLength: 120
          nullable: true
          description: >-
            Link or relink to an Athena (or other EHR) chart encounter. Pass
            null to clear (un-link). Accepts "Encounter/<id>" or bare "<id>".
...